idek anymore

+++ client/src/components/TripDetails.jsx
@@ -1,18 +1,31 @@
 import { Button } from '@mui/material';
+import useAuthContext from '../hooks/useAuthContext';
 
-function TripDetails({ trip }) {
+function TripDetails({ trip, setTrigger, trigger }) {
+
+    const { user } = useAuthContext();
 
     const handleClick = async () => {
+
+        if(!user){
+            return
+        }
+
         const response = await fetch(`http://localhost:5000/api/${trip._id}`, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json();
         console.log(json);
         if (response.ok) {
             console.log('trip deleted');
+            setTrigger(!trigger);
         } else {
             console.log('uhoh')
         }
+
     }
 
     return (
